WIP: Refactor a little further
